# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 給与明細・バックオフィス効率化システム API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: 社員一覧
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: 社員一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: 社員登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: 社員詳細
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: 社員詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: 社員更新
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: 社員削除（論理削除）
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── 月次実績 ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeCode:
          type: string
          description: 社員番号
        name:
          type: string
        nameKana:
          type: string
        department:
          type: string
          description: 部署
        position:
          type: string
          description: 役職
        baseSalary:
          type: number
          description: 基本給（円）
        transportationAllowance:
          type: number
          description: 通勤手当
        safetyDrivingAllowance:
          type: number
          description: 無事故手当
        longDistanceAllowance:
          type: number
          description: 長距離手当
        positionAllowance:
          type: number
          description: 役職手当
        familyAllowance:
          type: number
          description: 家族手当
        earlyOvertimeAllowance:
          type: number
          description: 早出残業手当
        commissionRatePerKm:
          type: number
          description: 歩合単価（円/km）
        commissionRatePerCase:
          type: number
          description: 歩合単価（円/件）
        dependentCount:
          type: integer
          description: 扶養親族数
        hasSpouse:
          type: boolean
          description: 配偶者の有無
        standardRemuneration:
          type: number
          description: 標準報酬月額（9月定時決定で確定、翌8月まで固定）
        careInsuranceApplied:
          type: boolean
          description: 介護保険適用（40〜64歳対象）
        healthInsuranceMonthly:
          type: number
          description: 健康保険料（月額）固定額
        pensionMonthly:
          type: number
          description: 厚生年金保険料（月額）固定額
        employmentInsuranceApplied:
          type: boolean
          description: 雇用保険適用
        residentTax:
          type: number
          description: 住民税（月額）
        salaryType:
          type: string
          description: 給与形態（fixed=固定給, daily=日給制）
          enum: [fixed, daily]
        hireDate:
          type: string
          format: date
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employeeCode
        - name
        - nameKana
        - department
        - baseSalary
        - transportationAllowance
        - safetyDrivingAllowance
        - longDistanceAllowance
        - positionAllowance
        - familyAllowance
        - earlyOvertimeAllowance
        - commissionRatePerKm
        - commissionRatePerCase
        - dependentCount
        - hasSpouse
        - standardRemuneration
        - careInsuranceApplied
        - healthInsuranceMonthly
        - pensionMonthly
        - employmentInsuranceApplied
        - residentTax
        - salaryType
        - hireDate
        - isActive
        - createdAt
        - updatedAt
    CreateEmployeeBody:
      type: object
      properties:
        employeeCode:
          type: string
        name:
          type: string
        nameKana:
          type: string
        department:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        transportationAllowance:
          type: number
        safetyDrivingAllowance:
          type: number
        longDistanceAllowance:
          type: number
        positionAllowance:
          type: number
        familyAllowance:
          type: number
        earlyOvertimeAllowance:
          type: number
        commissionRatePerKm:
          type: number
        commissionRatePerCase:
          type: number
        dependentCount:
          type: integer
        residentTax:
          type: number
        hireDate:
          type: string
          format: date
      required:
        - employeeCode
        - name
        - nameKana
        - department
        - baseSalary
        - dependentCount
        - hireDate
    UpdateEmployeeBody:
      type: object
      properties:
        employeeCode:
          type: string
        name:
          type: string
        nameKana:
          type: string
        department:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        transportationAllowance:
          type: number
        safetyDrivingAllowance:
          type: number
        longDistanceAllowance:
          type: number
        positionAllowance:
          type: number
        familyAllowance:
          type: number
        earlyOvertimeAllowance:
          type: number
        commissionRatePerKm:
          type: number
        commissionRatePerCase:
          type: number
        dependentCount:
          type: integer
        hasSpouse:
          type: boolean
        standardRemuneration:
          type: number
          description: 標準報酬月額（9月定時決定で確定、翌8月まで固定）
        careInsuranceApplied:
          type: boolean
          description: 介護保険適用（40〜64歳対象）
        healthInsuranceMonthly:
          type: number
        pensionMonthly:
          type: number
        employmentInsuranceApplied:
          type: boolean
        residentTax:
          type: number
        salaryType:
          type: string
          description: 給与形態（fixed=固定給, daily=日給制）
          enum: [fixed, daily]
        hireDate:
          type: string
          format: date
        isActive:
          type: boolean
...